requirement.txt
anyio==3.6.1
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
et-xmlfile==1.1.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
loguru==0.6.0
numpy==1.22.3
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pluggy==1.0.0
py==1.11.0
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
win32-setctime==1.1.0
wincertstore==0.2
XlsxWriter @ file:///opt/conda/conda-bld/xlsxwriter_1649073856329/work
